move func to parse_args because of slim
